[fix] RocketKV: int64 KT-cache block-index addressing (long-ctx/large-bs IMA)
The KT-cache update/read kernels compute the cache base address as
block_index * tokens_per_block * num_kv_heads * 2 * head_dim. Under the
V2-managed KT pool (Path A), KT shares the global page index, so
block_index = base_page * page_index_scale and spans the whole pool. At
long context + large batch the linear address exceeds INT32_MAX and wraps,
giving a CUDA illegal memory access. V1 kept KT in a dedicated compact
pool with small indices, so it never tripped this latent overflow.

Widen block_index to int64 before the multiply in the three KT kernels
(rocket_update_kt_cache_ctx/gen_kernel, rocket_paged_kt_cache_bmm_kernel).
Pure address-arithmetic widening: V1 small indices are numerically
unchanged. Matches the established pattern (dsa.py block_ids,
triton_prefill safe_page_ids, triton_paged_attention physical_page).

Verified on Llama-3.1-8B / B200: 16K input, batch 16, 1024 output tokens
runs clean (previously IMA at the first prefill); gsm8k 74.68 unchanged.

+++ tensorrt_llm/_torch/attention_backend/sparse/kernel.py
@@ -821,7 +821,7 @@ def rocket_update_kt_cache_gen_kernel(
                         batch_idx * max_kt_blocks_per_seq + block_offset_in_seq)
     token_idx_in_block = last_kt_token_idx % tokens_per_block
 
-    cache_base = ((block_idx * tokens_per_block + token_idx_in_block) *
+    cache_base = ((block_idx.to(tl.int64) * tokens_per_block + token_idx_in_block) *
                   num_kv_heads * 2 * head_dim + kv_head_idx * 2 * head_dim)
 
     cache_min_indices = cache_base + dim_indices
@@ -979,7 +979,7 @@ def rocket_update_kt_cache_ctx_kernel(
         tokens_in_block = kt_offsets % tokens_per_block
 
         # Calculate cache base addresses [BLOCK_SIZE_KT]
-        cache_bases = ((block_indices * tokens_per_block + tokens_in_block) *
+        cache_bases = ((block_indices.to(tl.int64) * tokens_per_block + tokens_in_block) *
                        num_kv_heads * 2 * head_dim + kv_head_idx * 2 * head_dim)
 
         cache_min_addrs = cache_bases[None, :] + dim_indices[:, None]
@@ -1131,7 +1131,7 @@ def rocket_paged_kt_cache_bmm_kernel(
         token_indices_in_block = kt_token_indices % tokens_per_block
 
         cache_bases = (
-            (block_indices * tokens_per_block + token_indices_in_block) *
+            (block_indices.to(tl.int64) * tokens_per_block + token_indices_in_block) *
             num_kv_heads * 2 * head_dim + kv_head_idx * 2 * head_dim)
 
         combined_mask = dim_mask[
